requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
numpy==2.4.0
pandas==2.3.3
scikit-learn==1.8.0
mlflow==3.9.0
shap==0.50.0
streamlit==1.54.0
python-dotenv==1.2.1
requests==2.32.5
joblib